+++ code/main-g474/app/bellow_curve.h
@@ -2,10 +2,9 @@
 #define APP_BELLOW_CURVE_H
 
 #include <stdint.h>
-#include "keyboard_layout.h"
 
 /* Cubic-Bezier response shape, reshaping an already-classified
- * bellow_classify() intensity (0..BELLOW_INTENSITY_MAX) through a curve from (0,0) to
+ * bellow_classify() intensity (0.0..1.0) through a curve from (0,0) to
  * (1,1) with two fully free 2D control points P1=(cx1,cy1), P2=(cx2,cy2),
  * each coordinate 0..256 (Q8, i.e. 0..1). This is the same construction
  * as CSS's cubic-bezier() timing function / cubic-bezier.com: dragging a
@@ -47,8 +46,8 @@
  * mathematical limit; in practice it lands within a fraction of a part in
  * 1024 of linear at the equivalent default (cx1=cy1=85, cx2=cy2=171, i.e.
  * both control points sitting on the diagonal), not a hard guarantee. */
-uint16_t bellow_curve_apply(uint16_t intensity,
-                             uint16_t cx1, uint16_t cy1,
-                             uint16_t cx2, uint16_t cy2);
+float bellow_curve_apply(float intensity,
+                         uint16_t cx1, uint16_t cy1,
+                         uint16_t cx2, uint16_t cy2);
 
 #endif /* APP_BELLOW_CURVE_H */
